requirements.txt
apscheduler>=3.10.4
beautifulsoup4>=4.13.3
colorlog>=6.9.0
discord-py>=2.5.2
google-genai>=1.0.0
python-dotenv>=1.0.0
pytz>=2025.2
requests>=2.32.3
tenacity>=8.2.0
tomli>=2.0.1; python_version < '3.11'
